pyproject.toml
[project]
name = "rtw"
version = "0.1.0"
description = "oneworld Explorer RTW ticket optimizer"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "haversine>=2.8",
    "airportsdata>=20240101",
    "fast-flights>=2.2",
    "playwright>=1.40",
    "keyring>=25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "hypothesis>=6.100",
    "pytest-recording>=0.13",
    "ruff>=0.5",
]

[project.scripts]
rtw = "rtw.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests requiring external services",
]
